backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
h11==0.16.0
idna==3.10
oauthlib==3.3.1
passlib==1.7.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
